Login, Act User,

+++ js/user.js
@@ -1,4 +1,6 @@
 let users = [];
+let actualUser = [];
+
 
 async function init(){
     loadUsers();
@@ -11,24 +13,33 @@ async function loadUsers(){
     }
 }
 async function register() {
-    registerBtn.disabled = true;
-    users.push({
-        email: email.value,
-        password: password.value,
-    });
-    await setItem('users', JSON.stringify(users));
-    resetForm();
+    let uname = document.getElementById('userName').value;
+    let uemail = document.getElementById('userEmail').value;
+    let upassword = document.getElementById('userpassword').value;
+    let upasswordconfirm = document.getElementById('userpasswordconfirm').value;
+    if (upassword == upasswordconfirm){
+        users.push({
+            name : uname,
+            email : uemail,
+            password : upassword,
+        })
+        await setItem('users', JSON.stringify(users));
+        resetForm();
+    }else{
+        alert('Passwörter stimmen nicht überein!')
+    }
+;
+
 }
 function resetForm() {
-    email.value = '';
-    password.value = '';
-    registerBtn.disabled = false;
+    windowlocation.href='Summary.html';
 }
 function loginUser(){
     let useremail = document.getElementById('userEmail');
     let userpassword = document.getElementById('userpassword');
     let user = users.find(u => u.email == useremail.value && u.password == userpassword.value)
     if(user){
+        localStorage.setItem('user', useremail.value);
         window.location.href='Summary.html'
     }else{
         alert('Bitte registrieren');
